+++ jest.config.js
@@ -49,9 +49,10 @@ module.exports = {
         testsForPackage('plugin-react-native-unhandled-rejection'),
         testsForPackage('plugin-react-native-hermes'),
         testsForPackage('plugin-react-native-client-sync'),
-        testsForPackage('plugin-react-native-global-error-handler'),
         testsForPackage('plugin-react-native-event-sync'),
+        testsForPackage('plugin-react-native-global-error-handler'),
         testsForPackage('plugin-react-native-session'),
+        testsForPackage('plugin-react-navigation'),
         testsForPackage('plugin-react-native-navigation')
       ],
       setupFiles: [